Add Usernotice functionality
In order to add USERNOTICE functionality, as well as to make future
message types more easy to add in the future line splitting and middle
parsing were changed.

Splitting the original line was moved into a dedicated function which
more cleanly handles the case of 2 line parts and 3 line parts for when
there's not a text portion and when there is.

Parsing the middle section was changed not to rely on strings.Split to
make it more versatile. This is done by using regex to extract the
explicit pieces we want. This also means that ROOMSTATE no longer needed
a special parse function.

Tests were also added for ensuring sub/resub/giftsub notices parse
correctly.